feat: migrate _request_uri to commom functions
